+++ R/mock-api.R
@@ -115,14 +115,16 @@ buildMockURL <- function (req, method="GET") {
     parts <- unlist(strsplit(url, "?", fixed=TRUE))
     ## Remove trailing slash
     f <- sub("\\/$", "", parts[1])
+    ## Sanitize the path to be portable for all R platforms
+    f <- gsub(":", "-", f)
     if (length(parts) > 1) {
         ## There's a query string. Append the digest as a suffix.
         f <- paste0(f, "-", hash(parts[2]))
     }
 
     ## Handle body and append its hash if present
-    if (length(body) > 0) {
-        f <- paste0(f, "-", hash(rawToChar(body)))
+    if (!is.null(body)) {
+        f <- paste0(f, "-", hash(body))
     }
 
     if (method == "DOWNLOAD") {
@@ -156,7 +158,23 @@ findMockFile <- function (file) {
     return(NULL)
 }
 
-requestBody <- function (req) req$options$postfields
+requestBody <- function (req) {
+    b <- req$options$postfields
+    if (length(b) > 0) {
+        ## Check length this way because b may be NULL or length 0 raw vector
+        b <- rawToChar(b)
+    } else {
+        b <- req$fields
+        if (!is.null(b)) {
+            ## Get a readable string representation
+            b <- deparse(b, control=NULL)
+            ## Strip out unhelpful indentation that it may add, then collapse
+            ## to single string, if broken into multiple lines
+            b <- paste(sub("^ +", "", b), collapse="")
+        }
+    }
+    return(b)
+}
 
 hash <- function (string, n=6) substr(digest(string), 1, n)