+++ tests/testthat/test-capture-requests.R
@@ -69,9 +69,11 @@ with_mock_API({
             readLines("api/object1.json"))
     })
 
-    test_that("Using simplify=FALSE", {
+    test_that("Using simplify=FALSE (and setting .mockPaths)", {
         d3 <- tempfile()
-        capture_requests(path=d3, simplify=FALSE, {
+        .mockPaths(d3)
+        on.exit(options(httptest.mock.paths=NULL))
+        capture_requests(simplify=FALSE, {
             GET("http://example.com/get/")
             utils::download.file("api/object1.json", tempfile())
         })
@@ -81,5 +83,23 @@ with_mock_API({
         expect_is(response, "response")
         expect_identical(content(response),
             content(GET("http://example.com/get/")))
+    })
+
+    test_that("Using verbose=TRUE (and .mockPaths)", {
+        d4 <- tempfile()
+        .mockPaths(d4)
+        on.exit(options(httptest.mock.paths=NULL))
+        capture_requests(verbose=TRUE, {
+            expect_message(GET("http://example.com/get/"),
+                "Writing .*example.com.get.json")
+            expect_message(utils::download.file("api/object1.json", tempfile()),
+                "Writing .*api.object1.json")
+        })
+        expect_true(setequal(dir(d4, recursive=TRUE),
+            c("example.com/get.json", "api/object1.json")))
+        expect_identical(readLines(file.path(d4, "example.com/get.json")),
+            readLines("example.com/get.json"))
+        expect_identical(readLines(file.path(d4, "api/object1.json")),
+            readLines("api/object1.json"))
     })
 })