+++ NEWS.md
@@ -1,3 +1,9 @@
+## httptest 2.1.0
+* `with_mock_API` and `without_internet` handle multipart and urlencoded form data in mocked HTTP requests.
+* `buildMockURL` escapes URL characters that are not valid in file names on all R platforms (which `R CMD check` would warn about).
+* `capture_requests` now has a `verbose` argument, which, if `TRUE`, prints a message with the file path where each captured request is written.
+* `capture_requests` takes the first element in `.mockPaths` as its default "path" argument. The default is unchanged since `.mockPaths` by default returns the current working directory, just as the "path" default was, but if you set a different mock path for reading mocks, `capture_requests` will write there as well.
+
 # httptest 2.0.0
 * `capture_requests` now writes non-JSON-content-type and non-200-status responses as full "response" objects in .R files. `with_mock_API` now looks for .R mocks if a .json mock isn't found. This allows all requests and all responses, not just JSON content, to be mocked.
 * New `.mockPaths` function, in the model of `.libPaths`, which allows you to specify alternate directories in which to search for mock API responses.